Introduce the new linter config to Expo SDK
Adds eslint-config-expo v6 to the Expo SDK. The rest of Universe stays the same but Expo SDK will use the new linter config. There's a good variety of editors people use to work on the SDK and contractors contribute to it as well, this is an aggressive litmus test.

Tested by opening expo-sdk in VS Code, linting and formatting work as I expect.

fbshipit-source-id: 77079ec