Add tests for sensors: Accelerometer, Magnetometers, Gyroscope
These are some example tests for the 3-axis device sensor modules. They use the mock functions on the native module mocks to verify calls to the native layer. I also exposed the NativeEventEmitter of the sensor wrapper to tests so we can emit events. This uses the new jest-expo which defines the magnetometer mocks.

Test plan: ran tests with yarn test, ensure they ran quickly

fbshipit-source-id: da521eaeab