Add exporting tests for Expo.js itself to catch missing `default`s
Brent caught a bug where we were missing a `default` when requiring a module that was using `export` syntax. Added a test that enumerates the properties of the Expo SDK and tests each one to make sure (a) we can import them and (b) they don't look like modules for which we're missing a `default`.

Test Plan: Ran unit tests in expo-sdk. Removed a default where we needed one and saw a test fail as expected.

fbshipit-source-id: 9353f04