Split logs into polyfill + serialization + transport with unit tests
Adds comprehensive unit tests for logs particularly because developers rely on logging when things go wrong. To aid with testability I split Logs.js into three parts that are tested independently as well as together.

In particular the tests simulate failures and concurrent logging calls (ex: logging while a prior batch of log messages is in transit to XDE) because failures will be hard to diagnose if we can't use logging to debug.

Test Plan: Run unit tests. Open NCL with Expo iOS and make sure the following are logged to exp the same way as before with symbolicated error stacks:
- unhandled promise rejections -> show rejection stack
- throwing an error -> show error and stack
- console.log(error) -> show error stack
- console.error(message) -> show stack of console.error call

fbshipit-source-id: 7e9271c