Use r-n-svg 5.5.1-exp.1 with `color` RGBA-parsing fix
Fixed a parsing bug in r-n-svg that affected colors with an alpha channel of 1.0 -- make it so we get back 1.0 instead of undefined. Tested with NCL on Android (no redbox) and NCL on iOS (didn't see excessive logs about getting a null value where a number was expected in Console.app).

fbshipit-source-id: 3ac5dc6